Last warning

+++ YasirRoBot/bot/plugins/ban_user.py
@@ -14,9 +14,9 @@ async def ban_handler(bot, m):
     user = m.command[1]
     if not await db.is_banned(user):
         await db.add_ban_user(user)
-        await m.reply(f"🚫 User {user} berhasil dibanned dari bot ini.")
+        await m.reply(f"🚫 User {user} succesfully banned from this bot.")
     else:
-        await m.reply(f"🚫 User {user} sudah dibanned sebelumnya.")
+        await m.reply(f"🚫 User {user} has banned forever before.")
 
 
 @StreamBot.on_message(filters.command("unban") & filters.private & filters.user(list(Var.OWNER_ID)))
@@ -25,5 +25,8 @@ async def rm_ban_handler(bot, m):
         await m.reply_text(text="🚫 Gunakan: /unban [id user] untuk unban pengguna.")
         return
     user = m.command[1]
-    await db.remove_ban(user)
-    await m.reply(f"🚫 User {user} berhasil diunbanned dari bot ini.")
+    if not await db.is_banned(user):
+        await m.reply(f"🚫 User {user} is not added banned list before.")
+    else:
+        await db.remove_ban(user)
+        await m.reply(f"🚫 User {user} has been banned from this bot.")
